+++ src/main/scala/edu/illinois/ncsa/fence/Server.scala
@@ -274,12 +274,6 @@ object Server extends TwitterServer {
     case (Get | Options, Root / "extractors" / "running") =>
       cf andThen tokenFilter andThen Clowder.clowderCatchAll(Path("/api/extractions/extractors_names"))
 
-    case (Get | Options, Root / "extractors" / "running" / "servers") =>
-      cf andThen tokenFilter andThen admin andThen Clowder.clowderCatchAll(Path("/api/extractions/servers_ips"))
-
-    case (Get | Options, Root / "extractors" / "instances") =>
-      cf andThen tokenFilter andThen admin andThen Clowder.clowderCatchAll(Path("/api/extractions/extractors_details"))
-
     case (Get | Options, Root / "extractors" / fileId / "status") =>
       cf andThen tokenFilter andThen new ResourceAuthFilter(fileId) andThen
         Clowder.clowderCatchAll(Path("/api/extractions/" + fileId + "/status"))