+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+## Removed
+ - Admin endpoints ```/extractors/running/servers``` and ```/extractors/instances``` [BD-2349](https://opensource.ncsa.illinois.edu/jira/browse/BD-2349) 
+
 ## [0.5.0] - 2018-10-04
 
 ## Added